fix(styles): Move custom styles to globals
