+++ src/styles/globals.css
@@ -52,6 +52,8 @@ code {
 }
 
 .header {
+    background-color: #5d8ea9;
+    color: white;
     padding-left: 3%;
     padding-right: 3%;
     padding-top: 1rem;
@@ -364,6 +366,7 @@ th {
 }
 
 .list-group-item-action {
+    background-color: white !important;
     color: #6b6c7e !important;
 }
 
@@ -443,6 +446,17 @@ th {
     border-color: #dbdce4 !important;
     color: black !important;
     background-color: #f8f9fa !important;
+}
+
+.sidebar-active {
+    border-color: #dbdce4 !important;
+    color: #000000 !important;
+    background-color: #e89734 !important;
+}
+
+.link-sidebar {
+    background-color: white !important;
+    color: #6b6c7e !important;
 }
 
 .label {
